+++ packages/cli/config/jest.js
@@ -54,14 +54,14 @@ function getRoleConfig(role) {
     case 'common-library':
     case 'frontend-plugin':
     case 'frontend-plugin-module':
-      return { testEnvironment: 'jsdom' };
+      return { testEnvironment: require.resolve('jest-environment-jsdom') };
     case 'cli':
     case 'backend':
     case 'node-library':
     case 'backend-plugin':
     case 'backend-plugin-module':
     default:
-      return { testEnvironment: 'node' };
+      return { testEnvironment: require.resolve('jest-environment-node') };
   }
 }