+++ packages/backend-defaults/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/backend-defaults
 
+## 0.1.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-app-api@0.2.3-next.1
+  - @backstage/backend-plugin-api@0.1.4-next.1
+
 ## 0.1.3-next.0
 
 ### Patch Changes