+++ packages/backend-tasks/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/backend-tasks
 
+## 0.3.7-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.0-next.1
+  - @backstage/config@1.0.4-next.0
+  - @backstage/errors@1.1.3-next.0
+  - @backstage/types@1.0.1-next.0
+
 ## 0.3.7-next.0
 
 ### Patch Changes