+++ packages/techdocs-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @techdocs/cli
 
+## 1.2.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.0-next.1
+  - @backstage/plugin-techdocs-node@1.4.2-next.1
+  - @backstage/catalog-model@1.1.3-next.0
+  - @backstage/cli-common@0.1.10
+  - @backstage/config@1.0.4-next.0
+
 ## 1.2.3-next.0
 
 ### Patch Changes