+++ plugins/code-coverage/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-code-coverage
 
+## 0.2.4-next.1
+
+### Patch Changes
+
+- fcab2579a0: Adds installation instructions
+- Updated dependencies
+  - @backstage/core-components@0.12.0-next.1
+  - @backstage/catalog-model@1.1.3-next.0
+  - @backstage/config@1.0.4-next.0
+  - @backstage/core-plugin-api@1.1.0-next.0
+  - @backstage/errors@1.1.3-next.0
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-react@1.2.1-next.1
+
 ## 0.2.4-next.0
 
 ### Patch Changes