+++ plugins/playlist/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-playlist
 
+## 0.1.2-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.0-next.1
+  - @backstage/catalog-model@1.1.3-next.0
+  - @backstage/core-plugin-api@1.1.0-next.0
+  - @backstage/errors@1.1.3-next.0
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-common@1.0.8-next.0
+  - @backstage/plugin-catalog-react@1.2.1-next.1
+  - @backstage/plugin-permission-common@0.7.1-next.0
+  - @backstage/plugin-permission-react@0.4.7-next.0
+  - @backstage/plugin-playlist-common@0.1.2-next.0
+  - @backstage/plugin-search-react@1.2.1-next.1
+
 ## 0.1.2-next.1
 
 ### Patch Changes