+++ plugins/search-backend-module-pg/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-search-backend-module-pg
 
+## 0.4.2-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.0-next.1
+  - @backstage/plugin-search-backend-node@1.0.4-next.1
+  - @backstage/config@1.0.4-next.0
+  - @backstage/plugin-search-common@1.1.1-next.0
+
 ## 0.4.2-next.0
 
 ### Patch Changes