+++ plugins/tech-insights-backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-tech-insights-backend
 
+## 0.5.4-next.1
+
+### Patch Changes
+
+- f12e9e5b8c: Add Documentation on 404 Errors
+- Updated dependencies
+  - @backstage/backend-common@0.16.0-next.1
+  - @backstage/backend-tasks@0.3.7-next.1
+  - @backstage/plugin-tech-insights-node@0.3.6-next.1
+  - @backstage/catalog-client@1.1.2-next.0
+  - @backstage/catalog-model@1.1.3-next.0
+  - @backstage/config@1.0.4-next.0
+  - @backstage/errors@1.1.3-next.0
+  - @backstage/types@1.0.1-next.0
+  - @backstage/plugin-tech-insights-common@0.2.8-next.0
+
 ## 0.5.4-next.0
 
 ### Patch Changes