+++ plugins/techdocs-node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-techdocs-node
 
+## 1.4.2-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.0-next.1
+  - @backstage/catalog-model@1.1.3-next.0
+  - @backstage/config@1.0.4-next.0
+  - @backstage/errors@1.1.3-next.0
+  - @backstage/integration@1.4.0-next.0
+  - @backstage/plugin-search-common@1.1.1-next.0
+
 ## 1.4.2-next.0
 
 ### Patch Changes